Make Compatibility check work in PHP 5.2
