+++ phpcs.xml.dist
@@ -58,6 +58,7 @@
 	<!-- The main plugin file should be parsable by PHP < 5.6. -->
 	<rule ref="Generic.Arrays.DisallowLongArraySyntax.Found">
 		<exclude-pattern>web-stories.php</exclude-pattern>
+		<exclude-pattern>Web_Stories_Compatibility.php</exclude-pattern>
 	</rule>
 
 	<!-- Show details about violated sniffs -->
